Downloads/DMS_OCR_MODEL/OCR/requirements.txt
# Core
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=1.10.7

# Azure
azure-storage-blob>=12.16.0
openai>=1.0.0

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.17.0
opencv-python-headless==4.8.0.74
tika>=3.1.0
python-magic==0.4.27
Pillow>=10.0.0
python-docx>=1.0.0
python-pptx>=0.6.21

# Data Processing
pandas>=2.0.0
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
numpy>=1.24.0

# HTTP & API
httpx>=0.24.1
aiohttp>=3.8.4
requests>=2.31.0

# Utils
tqdm>=4.65.0
python-magic-bin==0.4.14; platform_system == 'Windows'

# Optional: Uncomment if using Azure OpenAI
# openai>=1.0.0
# azure-identity>=1.13.0
# azure-keyvault-secrets>=4.6.0
